+++ Laptop/Robin.RetroEncabulator/MovementData.cs
@@ -2,9 +2,9 @@
 {
 	public struct MovementData
 	{
-		public int Direction { get; set; }
-		public int MoveSpeed { get; set; }
-		public int TurnSpeed { get; set; }
+		public short Direction { get; set; }
+		public short MoveSpeed { get; set; }
+		public short TurnSpeed { get; set; }
 		public bool DribblerEnabled { get; set; }
 
 		public int Duration { get; set; }
